requirements.txt
python-telegram-bot   13.13
selenium              4.4.3